+++ notes/Chord4List.cpp
@@ -23,16 +23,16 @@ Chord4List::Chord4List(const Options& options, int rt, bool show) {
             assert(chords.empty());
             return;
         }
-        chords.push_back(newChord);               // put a chord in the list
-        done = referenceChord.makeNext(options);  // advance to next chord
+        chords.push_back(newChord);               // Put a chord in the list.
+        done = referenceChord.makeNext(options);  // Advance to next chord.
     }
 
-    assert(!chords.empty());  // in theory ok, but don't know if we handle it.
+    assert(!chords.empty());  // In theory ok, but don't know if we handle it.
     std::sort(chords.begin(), chords.end(), [options](Chord4Ptr c1, Chord4Ptr c2) {
         return compareChords(options, c1, c2);
     });
 
-    // Now that we are sorted, add the ranks
+    // Now that we are sorted, add the ranks.
     int rank = 0;
     for (auto it : chords) {
         it->rank = rank++;