+++ tests/testHarmonyChords2023.cpp
@@ -121,7 +121,10 @@ static void testAnalyze56() {
     auto options = makeOptions(false);
     Chord4Manager mgr(options);
     Chord4Ptr chordA = Chord4::fromString(options, 5, "G1B2D3G3");
-    Chord4Ptr chordB = Chord4::fromString(options, 6, "E1C3E3A3");                                    // only one A - should be illegal
+    Chord4Ptr chordB = Chord4::fromString(options, 6, "E1C3E3A3"); 
+                                   // only one A - should be illegal
+    assert(chordA);
+    assert(chordB);
     testAnalyzeProgression(mgr, options, chordA, chordB, ProgressionAnalyzer::AVG_PENALTY_PER_RULE);  // this illegal
                                                                                                       // SQINFO("\n-------------------------bgf: end testAnalyze56");
 }
@@ -144,7 +147,7 @@ static void analyzeChordGenForParallelCase() {
     }
     SQINFO("\n\n---------------------- bgf: start analyzeChordGenForParallelCasex");
     auto options = makeOptions(false);
-    options.style->setRangesPreference(Style::Ranges::NARROW_RANGE);
+    options.style->setRangesPreference(Style::Ranges::VOCAL_RANGE);
     //   Chord4Ptr chordB = Chord4::fromString(options, 3, "E2B2E3G3");
     Chord4ListPtr list = std::make_shared<Chord4List>(options, 3, show);
     SQINFO("\n\n---------------------- bgf: end analyzeChordGenForParallelCasex");
@@ -156,7 +159,7 @@ static void analyzeChordGenForParallelCase() {
 static void analyzeReportedParallel() {
     SQINFO("\n\n---------------------- bgf: start analyzeReportedParallel");
     auto options = makeOptions(false);
-    options.style->setRangesPreference(Style::Ranges::NARROW_RANGE);
+    options.style->setRangesPreference(Style::Ranges::VOCAL_RANGE);
     Chord4Manager mgr(options);
     Chord4Ptr chordA = Chord4::fromString(options, 7, "D2B2D3F3");
     Chord4Ptr chordB = Chord4::fromString(options, 3, "E2B2E3G3");
@@ -181,7 +184,9 @@ static void analyzeReportedParallel() {
 void testHarmonyChords2023() {
     analyzeChordGenForParallelCase();
     canFindExpected56();
-    testAnalyze56();
+
+    SQINFO("put back testAnalyze56");
+    //testAnalyze56();
     testAnalyze56piston();
     testNoneInCommmon12();